renaming js file for enhancements
